+++ server/lib/intercom-client.js
@@ -1,12 +1,7 @@
 import request from "superagent";
 import prefixPlugin from "superagent-prefix";
 import superagentPromisePlugin from "superagent-promise-plugin";
-import JSONStream from "JSONStream";
-import tar from "tar-stream";
-import zlib from "zlib";
-import es from "event-stream";
 import _ from "lodash";
-import moment from "moment";
 
 export default class IntercomClient {
 
@@ -43,8 +38,13 @@ export default class IntercomClient {
         //   .diff(moment(), "seconds");
         // x-runtime
         this.instrumentationAgent.metricInc("api_call", 1, this.ship);
-        this.instrumentationAgent.metricVal("ratelimit_remaining", remaining, this.ship);
-        this.instrumentationAgent.metricVal("ratelimit_limit", limit, this.ship);
+        if (remaining) {
+          this.instrumentationAgent.metricVal("ratelimit_remaining", remaining, this.ship);
+        }
+
+        if (limit) {
+          this.instrumentationAgent.metricVal("ratelimit_limit", limit, this.ship);
+        }
       });
 
     if (this.accessToken) {
@@ -66,41 +66,6 @@ export default class IntercomClient {
   delete(url) {
     const req = this.req.delete(url);
     return this.attach(req);
-  }
-
-  /**
-   * Method to handle Mailchimp batch response as a JSON stream
-   * @param  {String} { response_body_url }
-   * @return {Stream}
-   */
-  handleResponse({ response_body_url }) {
-    const extract = tar.extract();
-    const decoder = JSONStream.parse();
-
-    extract.on("entry", (header, stream, callback) => {
-      if (header.name.match(/\.json/)) {
-        stream.pipe(decoder, { end: false });
-      }
-
-      stream.on("end", () => callback()); // ready for next entry
-      stream.on("error", () => callback()); // ready for next entry
-
-      stream.resume();
-    });
-
-    extract.on("finish", () => decoder.end());
-    extract.on("error", () => decoder.end());
-
-    request(response_body_url)
-      .pipe(zlib.createGunzip())
-      .pipe(extract);
-
-    return decoder
-      .pipe(es.map(function write(data, callback) {
-        return data.map(r => {
-          return callback(null, r);
-        });
-      }));
   }
 
   handleError(err) {