+++ server/lib/sync-agent/index.js
@@ -5,6 +5,7 @@ import Promise from "bluebird";
 import TagMapping from "./tag-mapping";
 import UserMapping from "./user-mapping";
 import WebhookAgent from "./webhook-agent";
+import EventsAgent from "./events-agent";
 
 export default class SyncAgent {
 
@@ -17,12 +18,17 @@ export default class SyncAgent {
     this.tagMapping = new TagMapping(intercomAgent, hullAgent, ship);
     this.userMapping = new UserMapping(ship);
     this.webhookAgent = new WebhookAgent(intercomAgent, hullAgent, ship, hostname);
+    this.eventsAgent = new EventsAgent(hullAgent, this.tagMapping, ship);
   }
 
   isConfigured() {
     return this.intercomAgent.intercomClient.ifConfigured();
   }
 
+  /**
+   * Makes sure that the Intercom account basic settings is in sync with Hull.
+   * That means tags to represent segments and webhook for events capturing.
+   */
   syncShip() {
     return this.webhookAgent.ensureWebhook()
       .then(() => this.hullAgent.getSegments())
@@ -58,7 +64,7 @@ export default class SyncAgent {
       const email = _.get(error, "data.email");
       const errorDetails = _.get(error, "error", []);
       const errorMessage = errorDetails.map(e => e.message).join(" ");
-      if (_.find(errorDetails, { code: "conflict"})) {
+      if (_.find(errorDetails, { code: "conflict" })) {
         return _.find(users, { email });
       }
 
@@ -85,10 +91,10 @@ export default class SyncAgent {
     return this.hullAgent.getSegments()
       .then(segments => {
         const ops = _.reduce(users, (o, user) => {
-          let userOp = {};
+          const userOp = {};
           if (!_.isEmpty(user["traits_intercom/id"])) {
             userOp.id = user["traits_intercom/id"];
-          } else if(!_.isEmpty(user.email)) {
+          } else if (!_.isEmpty(user.email)) {
             userOp.email = user.email;
           } else {
             return o;
@@ -100,7 +106,7 @@ export default class SyncAgent {
               return o;
             }
             o[segment.name] = o[segment.name] || [];
-            o[segment.name].push(userOp);
+            return o[segment.name].push(userOp);
           });
           user.remove_segment_ids.map(segment_id => {
             const segment = _.find(segments, { id: segment_id });
@@ -109,7 +115,7 @@ export default class SyncAgent {
               return o;
             }
             o[segment.name] = o[segment.name] || [];
-            o[segment.name].push(_.merge({}, userOp, {
+            return o[segment.name].push(_.merge({}, userOp, {
               untag: true
             }));
           });