+++ server/router/app.js
@@ -1,16 +1,23 @@
 import { Router } from "express";
 import bodyParser from "body-parser";
-import cors from 'cors';
+import cors from "cors";
+import { NotifHandler } from "hull";
 
-import ParseMessageMiddleware from "../util/middleware/parse-message";
-import NotifHandler from "../util/notif-handler";
-import ResponseMiddleware from "../util/middleware/response";
-import RequireConfiguration from "../util/middleware/require-configuration";
+import responseMiddleware from "../util/middleware/response";
+import requireConfiguration from "../util/middleware/require-configuration";
+import tokenMiddleware from "../util/middleware/token";
 
 export default function AppRouter(deps) {
   const router = new Router();
   const { hullMiddleware, appMiddleware } = deps;
   const { Actions, NotifHandlers } = deps.controllers;
+
+  const wrapWithMiddleware = (fn) => {
+    return (payload, context) => {
+      appMiddleware(context.req, {}, () => {});
+      return fn(payload, context);
+    };
+  };
 
   // FIXME: since we have two routers on the same mountpoint: "/"
   // all middleware applied here also is applied to the static router,
@@ -18,26 +25,28 @@ export default function AppRouter(deps) {
   // router.use(deps.hullMiddleware);
   // router.use(AppMiddleware(deps));
 
-  router.post("/fetch-all", hullMiddleware, appMiddleware, RequireConfiguration, Actions.fetchAll, ResponseMiddleware);
-  router.post("/batch", hullMiddleware, appMiddleware, RequireConfiguration, bodyParser.json(), Actions.batchHandler, ResponseMiddleware);
-  router.post("/notify", ParseMessageMiddleware, hullMiddleware, appMiddleware, NotifHandler(deps.Hull, {
+  const middlewareSet = [tokenMiddleware, hullMiddleware, appMiddleware, requireConfiguration, bodyParser.json()];
+
+  router.post("/fetch-all", ...middlewareSet, Actions.fetchAll, responseMiddleware);
+  router.post("/batch", ...middlewareSet, Actions.batchHandler, responseMiddleware);
+  router.post("/notify", NotifHandler({
     hostSecret: deps.shipConfig.hostSecret,
     groupTraits: false,
     handlers: {
-      "segment:update": NotifHandlers.segmentUpdateHandler,
-      "segment:delete": NotifHandlers.segmentDeleteHandler,
-      "user:update": NotifHandlers.userUpdateHandler,
-      "ship:update": NotifHandlers.shipUpdateHandler,
+      "segment:update": wrapWithMiddleware(NotifHandlers.segmentUpdateHandler),
+      "segment:delete": wrapWithMiddleware(NotifHandlers.segmentDeleteHandler),
+      "user:update": wrapWithMiddleware(NotifHandlers.userUpdateHandler),
+      "ship:update": wrapWithMiddleware(NotifHandlers.shipUpdateHandler)
     },
     shipCache: deps.shipCache
   }));
 
   // FIXME: 404 for that endpoint?
-  router.post("/intercom", hullMiddleware, appMiddleware, RequireConfiguration, bodyParser.json(), Actions.webhook, ResponseMiddleware);
+  router.post("/intercom", ...middlewareSet, Actions.webhook, responseMiddleware);
 
-  router.post("/sync", hullMiddleware, appMiddleware, RequireConfiguration, bodyParser.json(), Actions.sync, ResponseMiddleware);
+  router.post("/sync", ...middlewareSet, Actions.sync, responseMiddleware);
 
-  router.get("/schema/user_fields", cors(), hullMiddleware, appMiddleware, RequireConfiguration, bodyParser.json(), Actions.fields);
+  router.get("/schema/user_fields", cors(), ...middlewareSet, Actions.fields);
 
   return router;
 }