+++ server/util/instrumentation-agent.js
@@ -14,23 +14,25 @@ export default class InstrumentationAgent {
 
     if (process.env.SENTRY_URL) {
       console.log("starting raven");
-      this.raven = new raven.Client(process.env.SENTRY_URL);
+      this.raven = new raven.Client(process.env.SENTRY_URL, {
+        release: require(`${process.cwd()}/package.json`).version // eslint-disable-line global-require
+      });
       this.raven.patchGlobal();
     }
 
     if (process.env.LIBRATO_TOKEN && process.env.LIBRATO_USER) {
-     console.log("starting librato");
-     this.librato = require("librato-node"); // eslint-disable-line global-require
+      console.log("starting librato");
+      this.librato = require("librato-node"); // eslint-disable-line global-require
 
-     this.librato.configure({
-       email: process.env.LIBRATO_USER,
-       token: process.env.LIBRATO_TOKEN
-     });
-     this.librato.start();
-     this.librato.on("error", function onError(err) {
-       console.error("librato", err);
-     });
-   }
+      this.librato.configure({
+        email: process.env.LIBRATO_USER,
+        token: process.env.LIBRATO_TOKEN
+      });
+      this.librato.start();
+      this.librato.on("error", function onError(err) {
+        console.error("librato", err);
+      });
+    }
   }
 
   startTransaction(jobName, callback) {
@@ -78,5 +80,23 @@ export default class InstrumentationAgent {
     } catch (err) {
       console.warn("error in librato.measure", err);
     }
+  }
+
+  startMiddleware() {
+    if (this.raven) {
+      return raven.middleware.express.requestHandler(this.raven);
+    }
+    return (req, res, next) => {
+      next();
+    };
+  }
+
+  stopMiddleware() {
+    if (this.raven) {
+      return raven.middleware.express.errorHandler(this.raven);
+    }
+    return (req, res, next) => {
+      next();
+    };
   }
 }