+++ tests/intercom-client-tests.js
@@ -9,9 +9,14 @@ const shipMock = {
   }
 };
 
+const instrumentationAgent = {
+  metricInc: () => {},
+  metricVal: () => {}
+};
+
 describe("Intercom Client", () => {
   it("should allow for get api calls", () => {
-    const intercomClient = new IntercomClient(shipMock);
+    const intercomClient = new IntercomClient(shipMock, instrumentationAgent);
 
     return intercomClient.get("/users")
       .then(res => {
@@ -20,13 +25,13 @@ describe("Intercom Client", () => {
   });
 
   it("should provide error handling helper", () => {
-    const intercomClient = new IntercomClient(shipMock);
+    const intercomClient = new IntercomClient(shipMock, instrumentationAgent);
 
     return intercomClient.get("/non-existing-api")
       .catch(err => {
         const fErr = intercomClient.handleError(err);
-        assert.equal("404", fErr.extra.statusCode);
-        assert.equal("not_found", fErr.extra.body.errors[0].code);
+        assert.equal("404", fErr.statusCode);
+        assert.equal("not_found", fErr.body.errors[0].code);
       });
   });
 });