+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.0
+- capturing basic events from Intercom (skipping events for Hull Segments Tags)
+- adds `ensureWebhook` method which makes sure that the webhook has all topics set
+- fixes handling data coming from Hull and Intercom webhook
+- makes sure that companies, tags and segments array are clean 
+- overall maintenance fixes - linting, logging and cleanup
+
 ## 0.0.2
 - setting default `name` value for new users sent to Hull
 - minor bugfixes